replace all instances of uint with uint256

+++ contracts/lifecycle/Migrations.sol
@@ -8,9 +8,9 @@ import '../ownership/Ownable.sol';
  * @dev This is a truffle contract, needed for truffle integration, not meant for use by Zeppelin users.
  */
 contract Migrations is Ownable {
-  uint public lastCompletedMigration;
+  uint256 public lastCompletedMigration;
 
-  function setCompleted(uint completed) onlyOwner {
+  function setCompleted(uint256 completed) onlyOwner {
     lastCompletedMigration = completed;
   }
