+++ contracts/token/ERC20.sol
@@ -10,7 +10,7 @@ import './ERC20Basic.sol';
  */
 contract ERC20 is ERC20Basic {
   function allowance(address owner, address spender) constant returns (uint);
-  function transferFrom(address from, address to, uint value);
-  function approve(address spender, uint value);
-  event Approval(address indexed owner, address indexed spender, uint value);
+  function transferFrom(address from, address to, uint256 value);
+  function approve(address spender, uint256 value);
+  event Approval(address indexed owner, address indexed spender, uint256 value);
 }
