+++ test/helpers/BasicTokenMock.sol
@@ -7,7 +7,7 @@ import '../../contracts/token/BasicToken.sol';
 // mock class using BasicToken
 contract BasicTokenMock is BasicToken {
 
-  function BasicTokenMock(address initialAccount, uint initialBalance) {
+  function BasicTokenMock(address initialAccount, uint256 initialBalance) {
     balances[initialAccount] = initialBalance;
     totalSupply = initialBalance;
   }
